requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==5.1.5
django-crispy-forms==2.3
django-filter==24.3
djangorestframework==3.15.2
pillow==11.1.0
sqlparse==0.5.3
tzdata==2025.1
